+++ src/main.rs
@@ -54,6 +54,10 @@ struct Cli {
     /// Maximum number of tokens to use for each request
     max_tokens: u32,
 
+    #[arg(short = 'i', long)]
+    /// Maximum length of time in seconds to wait for an API request to complete
+    timeout: Option<u64>,
+
     #[arg(short = 'o', long, default_value_t = 1.0)]
     /// An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered. We generally recommend altering this or temperature but not both.
     top_p: f32,