+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add a separate flag for stdin, can now input text and have std at the same time
 - Add a flag for setting max tokens
+- Add timeout flag
 
 ## 2024-04-04 0.2.0